+++ Makefile
@@ -39,3 +39,5 @@ audit:
 format license-check lint test: ci
 	for x in $(pkgs); do (cd $$x && npm run $@) || exit 1; done
 
+doc:
+	cd lib && npm run doc
