+++ lib/package.json
@@ -39,6 +39,7 @@
   "scripts": {
     "build": "npm run clean && webpack && tsc -p tsconfig-esm.json && node prepareNodeESM.js && ../scripts/add-module-types.sh esm && ../scripts/fill-version.sh",
     "clean": "rm -rf {build,dist}",
+    "doc": "typedoc --out dist/docs src/index.ts",
     "format": "prettier --write \"{src,test}/**/*.ts\"",
     "license-check": "license-checker --production --onlyAllow 'Apache-2.0; BSD; ISC; MIT'",
     "lint": "eslint ./{src,test}/**/*.ts",
@@ -74,7 +75,8 @@
     "rollup": "^2.70.1",
     "sinon": "~13.0.1",
     "ts-loader": "^9.2.8",
-    "typescript": "^4.5.2",
+    "typedoc": "^0.22.15",
+    "typescript": "4.6.3",
     "webpack": "^5.72.0",
     "webpack-cli": "^4.9.2"
   }
