--- lib/src/nanotdf/types/CurveNameType.ts
@@ -1,6 +1,0 @@
-type CurveNameType = {
-  name: string;
-  length: number;
-};
-
-export default CurveNameType;
