--- lib/src/nanotdf/types/TypedArray.ts
@@ -1,7 +1,0 @@
-type TypedArray = {
-  BYTES_PER_ELEMENT: number;
-  set(array: ArrayLike<number>, offset?: number): void;
-  slice(start?: number, end?: number): TypedArray;
-};
-
-export default TypedArray;
