--- lib/src/nanotdf/utils/index.d.ts
@@ -1,1 +1,0 @@
-declare module './utils/utils';
