+++ lib/src/version.ts
@@ -1,2 +1,9 @@
+/**
+ * Exposes the released version number of the `@opentdf/client` package
+ */
 export const version = process.env.PKG_VERSION;
+
+/**
+ * A string name used to label requests as coming from this library client.
+ */
 export const clientType = 'client-web';
