+++ lib/test/nanotdf/Client.test.ts
@@ -1,6 +1,7 @@
 import { expect } from '@esm-bundle/chai';
+import { clientAuthProvider } from '../../src/auth/providers.js';
+import Client from '../../src/nanotdf/Client.js';
 
-import Client, { clientAuthProvider } from '../../src/nanotdf/Client.js';
 
 describe('nanotdf client', () => {
   it('Can create a client with a mock EAS', async () => {
